feature/PdfBuilder: optimized PdfCreationBuilder
