+++ src/main/java/edu/remad/tutoring2/services/pdf/PDFComplexInvoiceBuilder.java
@@ -21,7 +21,6 @@ import edu.remad.tutoring2.services.pdf.constants.ContentLayoutDataConstants;
 import edu.remad.tutoring2.services.pdf.documentinformation.DocumentInformationBuilder;
 import edu.remad.tutoring2.services.pdf.documentinformation.DocumentInformationUtilities;
 import edu.remad.tutoring2.services.pdf.exceptions.PDFComplexInvoiceBuilderException;
-import edu.remad.tutoring2.services.pdf.pagecontent.PDFCreationBuilder;
 import edu.remad.tutoring2.services.pdf.pagecontent.SinglePageContentLayouter;
 import edu.remad.tutoring2.services.pdf.utilities.PdfUtilities;
